Updated all read/write operations to happen with 16-bit values.
Even though the documentation says that most of the registers on the
v812 module are 8 bits wide, apparently they all still need to be
written with 16-bit values. Which, looking at the spacing of the
VME addresses of the module, does make sense...

So I updated all values to be written using uint16_t values. At the
same time I also added a bit of code for reading the module identifiers
as a first step of the configuration. Just to convince myself that
the code was really communicating with the device as I would've
expected it to.